feat(auditActions): Refactor audits handlers according to convention in order to be used in th-datat
Mainly change  the function names and update the query object interface
